// src/types/mock-data.type.ts
import { UserType } from './user.type.js';

export type MockData = {
  name: string[];
  description: string[];
  genre: string[];
  released: string[];
  previewLink: string[];
  videoLink: string[];
  starring: string[];
  director: string[];
  users: UserType[];
  posterImage: string[];
  backgroundImage: string[];
  backgroundColor: string[];
}